feat #3060: Added confirmation dialog for email deletion in EmailPreviewModal Added a confirmation dialog for unsaved emails in EmailPreviewModal to prevent unintentional email losses. A dialog appears prompting the user to either discard the email or continue editing when attempting to close the modal with unsaved changes. Additionally, updated the tooltip's label for copying email link. refactor: improve structure and readability of compose email related components
